auth-service/requirements.txt
redis==5.0.3
fastapi==0.109.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
uvicorn==0.29.0
gunicorn==22.0.0
SQLAlchemy==2.0.30
orjson==3.10.0
python-multipart==0.0.9
shortuuid==1.0.13
aiohttp==3.9.3
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
typer==0.12.3
async-fastapi-jwt-auth[asymmetric]==0.6.5
Werkzeug==3.0.2
alembic==1.13.1
psycopg2-binary==2.9.9
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
python-dateutil==2.9.0.post0
